+++ CHANGELOG.md
@@ -5,11 +5,21 @@ All notable changes to the "wevis-intellisense" extension will be documented in 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.2.0] — 2022-11-21
+## [1.2.0] — 2022-11-22
 
 ### Added
 
 - Snippets for all components; available for React.js, Vue.js, and Svelte.
+
+### Changed
+
+- Merge settings `wevisIntellisense.htmlLanguages` and `wevisIntellisense.javascriptLanguages` into `wevisIntellisense.autosuggestLanguages`.
+- Unified class matching pattern.
+
+### Fixed
+
+- Refactor autosuggestion functions.
+- Refactor `activate` function and disposing workspace events.
 
 ## [1.1.1] — 2022-11-19